+++ torch/sparse/nn/__init__.py
@@ -1,0 +1,2 @@
+import torch
+from . import functional
